feat: restructure layout component for locale support and improved internationalization
