+++ app/[locale]/layout.tsx
@@ -1,0 +1,57 @@
+import type { Metadata } from "next";
+import { Geist, Geist_Mono } from "next/font/google";
+import { NextIntlClientProvider, hasLocale } from "next-intl";
+import { notFound } from "next/navigation";
+import { routing } from "@/i18n/routing";
+import "./globals.css";
+
+// 1. Importera Toaster från sonner
+import { Toaster } from "sonner";
+import NavBar from "@/components/navbar/Navbar";
+import { getServerAuthProvider } from "@/lib/auth/factory/getServerProvider";
+import Footer from "@/components/Footer";
+import { getTranslations } from "next-intl/server";
+
+// Dina valfria Google-fonts
+const geistSans = Geist({ variable: "--font-geist-sans", subsets: ["latin"] });
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+});
+
+export const metadata: Metadata = {
+  title: "MindSwarm Quizes",
+  description: "Generated by create next app",
+};
+export default async function RootLayout({
+  children,
+  params,
+}: {
+  children: React.ReactNode;
+  params: { locale: string };
+}) {
+  const { locale } = await params;
+
+  if (!hasLocale(routing.locales, locale)) {
+    notFound();
+  }
+
+  const t = await getTranslations();
+  const authProvider = getServerAuthProvider();
+  const user = await authProvider.getCurrentUser();
+
+  return (
+    <html lang={locale}>
+      <body
+        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+      >
+        <NextIntlClientProvider locale={locale}>
+          <Toaster position="bottom-right" />
+          <NavBar title={t("title")} logo={t("logo")} user={user} />
+          {children}
+          <Footer title={t("title")} />
+        </NextIntlClientProvider>
+      </body>
+    </html>
+  );
+}
