--- app/layout.tsx
@@ -1,42 +1,0 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import siteConfig from "@/siteConfig";
-import "./globals.css";
-
-// 1. Importera Toaster från sonner
-import { Toaster } from "sonner";
-import NavBar from "@/components/navbar/Navbar";
-import { getServerAuthProvider } from "@/lib/auth/factory/getServerProvider";
-import Footer from "@/components/Footer";
-
-// Dina valfria Google-fonts
-const geistSans = Geist({ variable: "--font-geist-sans", subsets: ["latin"] });
-const geistMono = Geist_Mono({ variable: "--font-geist-mono", subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "MindSwarm Quizes",
-  description: "Generated by create next app",
-};
-
-export default async function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const authProvider = getServerAuthProvider();
-  const user = await authProvider.getCurrentUser();
-  return (
-    <html lang="en">
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
-        <Toaster position="bottom-right" />
-        <NavBar
-          title={siteConfig.title}
-          logo={siteConfig.logo?? ""}
-          user={user}
-        />
-        {children}
-        <Footer title={siteConfig.title} />
-      </body>
-    </html>
-  );
-}
